+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 添加 API `/api/filetree/removeDoc` | Add API `/api/filetree/removeDoc`
 - 添加 API `/api/filetree/getHPathByPath` | Add API `/api/filetree/getHPathByPath`
 - 添加 API `/api/filetree/getHPathByID` | Add API `/api/filetree/getHPathByID`
+- 添加 API `/api/filetree/moveDocs` | Add API `/api/filetree/moveDocs`
 
 ## v0.0.5 / 2023-06-30